Se agrega redondeo de bigDecimal a dos digitos decimales
Se cambia valor de cantidad por lote en pedidosManuales para poder ingresar una cantidad mas grande y con numeros decimales

+++ zapateria/sql/scriptBD.sql
@@ -201,7 +201,7 @@ CREATE TABLE `productosDeProveedor`
     `IdProveedor` int(11) NOT NULL,
     `IdMaestroProducto` int(11) NOT NULL,
     `PrecioVenta` double(45,2) NOT NULL,
-    `CantidadPorLote` int(11) NOT NULL,
+    `CantidadPorLote` double(45,2) NOT NULL,
     PRIMARY KEY (`Id`)
 );
